% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_posterior.R
\name{phi_posterior}
\alias{phi_posterior}
\title{posterior distribution}
\usage{
phi_posterior(y, lambda, aa, bb, R_ante)
}
\arguments{
\item{y}{Your name (character string; e.g. "john doe").}

\item{lambda}{Your name (character string; e.g. "john doe").}

\item{aa}{Your name (character string; e.g. "john doe").}

\item{bb}{Your name (character string; e.g. "john doe").}

\item{R_ante}{Your name (character string; e.g. "john doe").}
}
\description{
Greet a person and appropriately capitalize their name.
}
\details{
Function to take a random draw from the p(phi_j | ....)
}
