% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphasetype.R
\name{rPhDistribution}
\alias{rPhDistribution}
\title{Markov Chain Monte Carlo phase type parameters estimation}
\usage{
rPhDistribution(n, alpha = alpha, T = T)
}
\arguments{
\item{n}{Your name (character string; e.g. "john doe").}

\item{alpha}{Your name (character string; e.g. "john doe").}

\item{T}{Your name (character string; e.g. "john doe").}
}
\value{
A Markov Chain Monte Carlo
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
library(bnpPhaseType)

}
