% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Scores of Group A and Group B}
\format{
A data frame with 60 rows and 2 variables:
\describe{
  \item{X}{simulated values}
}
}
\source{
<https://www.github.com/mvuorre/exampleRPackage>
}
\usage{
simdata
}
\description{
A data set with the scores of two groups.
}
\keyword{datasets}
