% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights}
\alias{weights}
\title{function to generate the weights using the stick-breaking representation}
\usage{
weights(N, v)
}
\arguments{
\item{N}{Your name (character string; e.g. "john doe").}

\item{v}{Your name (character string; e.g. "john doe").}
}
\value{
A weights vector with the Stick-Breaking representation
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
