% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcErlangMix.R
\name{mcmcErlangMix}
\alias{mcmcErlangMix}
\title{Markov Chain Monte Carlo phase type parameters estimation}
\usage{
mcmcErlangMix(Y, a, b, aa, bb, alpha, beta, nscan, nburn, nskip)
}
\arguments{
\item{Y}{Your name (character string; e.g. "john doe").}

\item{a}{Your name (character string; e.g. "john doe").}

\item{b}{Your name (character string; e.g. "john doe").}

\item{aa}{Your name (character string; e.g. "john doe").}

\item{bb}{Your name (character string; e.g. "john doe").}

\item{alpha}{Your name (character string; e.g. "john doe").}

\item{beta}{Your name (character string; e.g. "john doe").}

\item{nscan}{Your name (character string; e.g. "john doe").}

\item{nburn}{Your name (character string; e.g. "john  doe").}

\item{nskip}{Your name (character string; e.g. "john doe").}
}
\value{
A Markov Chain Monte Carlo
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
library(bnpPhaseType)

data("simdata")
y = simdata$X
hist(y, breaks = 50)
fys = mcmcErlangMix( Y=y, a=0.1, b=0.1, aa=2, bb=0.1,
                   alpha=1, beta=1, nscan=10000, nburn=2000, nskip=8 )
}
