% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcErlangMix.R
\name{mcmcErlangMix}
\alias{mcmcErlangMix}
\title{Markov Chain Monte Carlo phase type parameters estimation}
\usage{
mcmcErlangMix(Y, a, b, aa, bb, alpha, beta, nscan, nburn, nskip)
}
\arguments{
\item{Y}{Your name (character string; e.g. "john doe").}

\item{a}{Your name (character string; e.g. "john doe").}

\item{b}{Your name (character string; e.g. "john doe").}

\item{aa}{Your name (character string; e.g. "john doe").}

\item{bb}{Your name (character string; e.g. "john doe").}

\item{alpha}{Your name (character string; e.g. "john doe").}

\item{beta}{Your name (character string; e.g. "john doe").}

\item{nscan}{Your name (character string; e.g. "john doe").}

\item{nburn}{Your name (character string; e.g. "john doe").}

\item{nskip}{Your name (character string; e.g. "john doe").}
}
\value{
A Markov Chain Monte Carlo
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
library(GeneralizedHyperbolic)
library(bnpPhaseType)

data("simdata")
y = simdata$X
hist(y, breaks = 50)
# Libraries 
fys = mcmcErlangMix( Y=y, a=0.1, b=0.1, aa=2, bb=0.1, alpha=1, beta=1, nscan=10000, nburn=2000, nskip=8 )
# Grid
tau.grid = seq( 0, 1.2*max(y), 0.1 )
# True density
p1=0.7 
p2=0.3
fyr=p1*dgig(tau.grid, chi = 1, psi = 2, lambda = 12.0)+p2*dgig(tau.grid, chi = 1, psi = 2, lambda = 30.0)
p2.5=function(x){quantile(x, probs=0.025)}
p97.5=function(x){quantile(x, probs=0.975)}
meanfy=apply(fys, 2, mean)
linffy=apply(fys, 2, p2.5)
lsupfy=apply(fys, 2, p97.5)
plot(tau.grid, meanfy, type="l", ylim=c(0,1.2*max(fyr)), xlab="y", ylab="fy")
polygon(x=c(tau.grid,rev(tau.grid)), y=c(linffy,rev(lsupfy)), lty=1, density=-1, col="grey", border="grey")
polygon(x=c(tau.grid,rev(tau.grid)), y=c(linffy,rev(lsupfy)), lty=1, density=-1, col="grey", border="grey")
lines(tau.grid, meanfy, lty=2)
lines(tau.grid, fyr, lty=1)

}
